Refactor number translator definition into a generator method
